% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dataframe.R
\name{plot_dataframe}
\alias{plot_dataframe}
\title{Plot summaries for a dataframe}
\usage{
plot_dataframe(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A list of ggplot objects, one per variable.
}
\description{
Creates histograms for numeric variables and bar charts for categorical variables.
}
\examples{
plot_dataframe(iris)
}
