% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical.R
\name{summarize_categorical}
\alias{summarize_categorical}
\title{Summarize categorical variables}
\usage{
summarize_categorical(data, var)
}
\arguments{
\item{data}{A data frame.}

\item{var}{A column name (string) referring to a categorical variable.}
}
\value{
A tibble with counts and proportions.
}
\description{
This function calculates counts and proportions for a categorical column in a data frame.
}
\examples{
summarize_categorical(iris, "Species")
}
