% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_dataframe.R
\name{summarize_dataframe}
\alias{summarize_dataframe}
\title{Summarize all columns in a data frame}
\usage{
summarize_dataframe(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A list with two tibbles: numeric and categorical summaries.
}
\description{
This function applies numeric and categorical summaries to each column in a data frame.
}
\examples{
summarize_dataframe(iris)
}
