% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_numeric.R
\name{summarize_numeric}
\alias{summarize_numeric}
\title{Summarize numeric variables}
\usage{
summarize_numeric(data, var)
}
\arguments{
\item{data}{A data frame.}

\item{var}{A column name (string) referring to a numeric variable.}
}
\value{
A tibble with summary statistics.
}
\description{
This function calculates summary statistics for a numeric column in a data frame.
}
\examples{
summarize_numeric(iris, "Sepal.Length")
}
